src/app/requirements.txt
streamlit
facenet-pytorch
torch
torchvision
Pillow
qdrant-client
python-dotenv
datasets
tqdm
numpy
opencv-python
huggingface-hub
requests
urllib3
blinker
